+++ tests/unit/documentation-helpers.spec.ts
@@ -816,7 +816,7 @@ describe("Documentation helpers", () => {
   });
 
   describe("depictSecurity()", () => {
-    test("should handle Basic, Bearer and CustomHeader Securities", () => {
+    test("should handle Basic, Bearer and Header Securities", () => {
       expect(
         depictSecurity([
           [{ type: "basic" }, { type: "bearer" }],