+++ tests/unit/migration.spec.ts
@@ -1,17 +1,15 @@
 import { RuleTester } from "@typescript-eslint/rule-tester";
 import migration from "../../src/migration";
-// import parser from "@typescript-eslint/parser";
+import parser from "@typescript-eslint/parser";
 import { version } from "../../package.json";
 
 RuleTester.afterAll = afterAll;
 RuleTester.describe = describe;
 RuleTester.it = it;
 
-/*
 const tester = new RuleTester({
   languageOptions: { parser },
 });
-*/
 
 describe("Migration", () => {
   test("should consist of one rule being the major version of the package", () => {
@@ -19,10 +17,23 @@ describe("Migration", () => {
     expect(migration).toMatchSnapshot();
   });
 
-  /*
   tester.run("v23", migration.rules.v23, {
-    valid: [],
-    invalid: [],
+    valid: [`import { HeaderSecurity } from "express-zod-api";`],
+    invalid: [
+      {
+        code: `const security: CustomHeaderSecurity = {};`,
+        output: `const security: HeaderSecurity = {};`,
+        errors: [
+          {
+            messageId: "change",
+            data: {
+              subject: "interface",
+              from: "CustomHeaderSecurity",
+              to: "HeaderSecurity",
+            },
+          },
+        ],
+      },
+    ],
   });
-  */
 });