+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## Version 23
+
+### v23.0.0
+
+- Publicly exposed interface `CustomHeaderSecurity` is renamed to `HeaderSecurity`;
+- Restrictions have been placed between header names in the middleware `security` declaration and in `input` properties.
+
 ## Version 22
 
 ### v22.4.1